+++ src/xrEngine/line_edit_control.cpp
@@ -637,7 +637,7 @@ void line_edit_control::move_pos_left()
 void line_edit_control::move_pos_right() { ++m_cur_pos; }
 void line_edit_control::move_pos_left_word()
 {
-    size_t i = m_cur_pos > 0 ? m_cur_pos - 1 : 0;
+    int i = m_cur_pos > 0 ? m_cur_pos - 1 : 0;
 
     while (i > 0 && m_edit_str[i] == ' ')
         --i;
